hope-ecommerce/pyproject.toml
[tool.poetry]
name = "hope-ecommerce"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
uvicorn = "^0.27.1"
firebase-admin = "^6.4.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
